Feat: Board 생성, 수정 유효성 검사

+++ src/main/java/pairproject/foodmap/service/BoardImageService.java
@@ -42,11 +42,13 @@ public class BoardImageService {
                                              List<MultipartFile> addFiles,
                                              List<String> deleteFilenames) {
         fileUtil.updateFileDir(fileDir);
-        fileUtil.deleteAll(deleteFilenames);
         List<String> addFilenames = fileUtil.saveAll(addFiles);
-
         boardImageMapper.save(addFilenames, boardId);
-        boardImageMapper.delete(deleteFilenames);
+
+        if (deleteFilenames != null) {
+            fileUtil.deleteAll(deleteFilenames);
+            boardImageMapper.delete(deleteFilenames);
+        }
         return getBoardImageAllByBoardId(boardId);
     }
